Add the 'when' parameter to io.color()
This parameter allows color to be turned off based on the TTY python is
attached to.  This parameter used to be in write_color(), but I moved it
to color() because sometimes programs want to build up strings before
printing them and still need to be able to use this functionality.