Report correct line numbers.
Before this commit, the logging helpers in debug.py were not reporting
the correct line numbers.  The problem is that the logging module
attaches line numbers to all log messages behind the scenes, and of
course it was attaching the line numbers from the debug module itself.
To get the right numbers to be attached, I had to monkey-patch the
logging module to trick it into inspecting the right frame.  The
monkey-patching undone immediately after the message is sent, so nothing
outside this module should be affected.